Change int->long for sales

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -3,18 +3,18 @@ package ru.netology.stats;
 public class StatsService {
 
     // #1 Метод возвращающий сумму всех продаж
-    public int sumOfAllSales(int[] sales) {
-        int sumMonth = 0; // изначальное значение суммы продаж за все месяцы
+    public int sumOfAllSales(long[] sales) {
+        long sumMonth = 0; // изначальное значение суммы продаж за все месяцы
 
         for (int i = 0; i < sales.length; i++) {
             sumMonth = sumMonth + sales[i];
         }
 
-        return sumMonth; // сума значений всех элементов массива
+        return (int) sumMonth; // сума значений всех элементов массива
     }
 
     // #2 Метод возвращающий среднюю сумму продаж в месяц
-    public int avrOfAllSales(int[] sales) {
+    public int avrOfAllSales(long[] sales) {
         int avrMonth; // объявление переменной для средней суммы продаж
         int sumMonth = sumOfAllSales(sales); // вызов метода #1 возвращающего сумму всех продаж
         avrMonth = sumMonth / sales.length; // расчет средней суммы продаж в месяц
@@ -23,7 +23,7 @@ public class StatsService {
     }
 
     // #3 Метод возвращающий номер месяца, в котором был пик продаж, то есть осуществлены продажи на максимальную сумму
-    public int numberOfMonthWithMaxSales(int[] sales) {
+    public int numberOfMonthWithMaxSales(long[] sales) {
         int maxMonth = 0; // номер месяца с максимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -36,7 +36,7 @@ public class StatsService {
     }
 
     // #4 Метод возвращающий номер месяца с минимальными продажами
-    public int minSales(int[] sales) {
+    public int minSales(long[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -49,7 +49,7 @@ public class StatsService {
     }
 
     // #5 Метод возвращающий количество месяцев, в которых продажи были ниже среднего
-    public int countOfMonthWithLessAvrSales(int[] sales) {
+    public int countOfMonthWithLessAvrSales(long[] sales) {
         int lessMonth = 0; // число месяцев с продажами ниже среднего значения
         int avrMonth = avrOfAllSales(sales); // вызов метода #2 метода возвращающего среднюю сумму продаж в месяц
 
@@ -63,7 +63,7 @@ public class StatsService {
     }
 
     // #6 Метод возвращающий количество месяцев, в которых продажи были выше среднего
-    public int countOfMonthWithMoreAvrSales(int[] sales) {
+    public int countOfMonthWithMoreAvrSales(long[] sales) {
         int moreMonth = 0; // число месяцев с продажами выше среднего значения
         int avrMonth = avrOfAllSales(sales); // вызов метода #2 метода возвращающего среднюю сумму продаж в месяц
